Refactor: 기존에 탑레벨 클래스 Operator 삭제 -> 중첩 클래스로 추가
- 현재 한 곳에서만 사용되는 Operator 클래스를 중첩 클래스로 변경하였음
- ArithmeticLogicUnitUsingStack 이 인터페이스 이기 때문에 다른 구현체가 더 존재할 경우, 산술연산과 관련해서는 같은 로직을 사용할 것 이기에, 그 때는 기존처럼 Operator 를 탑레벨 클래스로 뽑아내는 게 좋을 듯 하다.
- 현재로서는 하나의 클래스에서만 사용되는 Operator 를 외부로부터 감추고, 사용하는 클래스와 가깝게 위치시는게 더 좋을 것 이라 생각함
